+++ software/scripts/synle-make-component.js
@@ -6,12 +6,16 @@ async function doWork() {
 
   console.log("  >> Download and installing synle-make-component:", consoleLogColor4(targetPath));
 
+  if (TEST_FORCE_REFRESH) {
+    await deleteFolder(targetPath);
+  }
+
+  exitIfPathFound(targetPath);
+
   console.log("    >> Cloning and building", consoleLogColor4(targetPath));
 
-  await deleteFolder(targetPath);
-
   await execBash(`
-    git clone --depth 1 -b master https://github.com/synle/make-component.git "${targetPath}"
+    git clone --depth 1 --single-branch -b master https://github.com/synle/make-component.git "${targetPath}"
   `);
 
   await execBash(`npm i && npm run build`, false, {
@@ -25,13 +29,13 @@ async function doWork() {
       [ -f ${targetPath}/setup.sh ] && . ${targetPath}/setup.sh
 
       ## fzf alias for synle-make-component
-      getMakeComponentOptions(){
+      get_make_component_options(){
         make-help
       }
 
-      fuzzyMakeComponent(){
+      fuzzy_make_component(){
         makeComponentCommand=$(( \
-        getMakeComponentOptions \
+        get_make_component_options \
         ) | sed '/^\s*$/d' | uniq | fzf)
         echo "$makeComponentCommand"
         $makeComponentCommand
